pyproject.toml
[project]
name = "ga-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.40.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.3",
    "types-requests[dev]>=2.32.0.20250328",
]

[project.optional-dependencies]
dev = [
    "google-auth-stubs>=0.3.0",
    "python-dotenv>=1.1.0",
    "types-requests>=2.32.0.20250328",
]
